celery_queue/requirements.txt
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.3
flower==2.0.1
requests==2.31.0
pydub==0.25.1
boto3==1.34.77
opencv-python==4.9.0.80
numpy==1.26.4
